Refactor: Move poroelast scatra input to physics
