+++ src/global_legacy_module/4C_global_legacy_module_validparameters.cpp
@@ -41,7 +41,6 @@
 #include "4C_inpar_particle.hpp"
 #include "4C_inpar_pasi.hpp"
 #include "4C_inpar_plasticity.hpp"
-#include "4C_inpar_poroscatra.hpp"
 #include "4C_inpar_problemtype.hpp"
 #include "4C_inpar_s2i.hpp"
 #include "4C_inpar_scatra.hpp"
@@ -61,6 +60,7 @@
 #include "4C_linear_solver_method_input.hpp"
 #include "4C_lubrication_input.hpp"
 #include "4C_poroelast_input.hpp"
+#include "4C_poroelast_scatra_input.hpp"
 #include "4C_porofluid_pressure_based_elast_input.hpp"
 #include "4C_porofluid_pressure_based_elast_scatra_input.hpp"
 #include "4C_porofluid_pressure_based_input.hpp"
@@ -253,7 +253,7 @@ std::map<std::string, Core::IO::InputSpec> Global::valid_parameters()
   Inpar::S2I::set_valid_parameters(specs);
   Inpar::FS3I::set_valid_parameters(specs);
   PoroElast::set_valid_parameters(specs);
-  Inpar::PoroScaTra::set_valid_parameters(specs);
+  PoroElastScaTra::set_valid_parameters(specs);
   PoroPressureBased::set_valid_parameters_porofluid(specs);
   PoroPressureBased::set_valid_parameters_porofluid_elast_scatra(specs);
   PoroPressureBased::set_valid_parameters_porofluid_elast(specs);
