+++ src/poroelast_scatra/4C_poroelast_scatra_input.cpp
@@ -5,7 +5,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "4C_inpar_poroscatra.hpp"
+#include "4C_poroelast_scatra_input.hpp"
 
 #include "4C_inpar_scatra.hpp"
 #include "4C_io_input_spec_builders.hpp"
@@ -14,7 +14,7 @@ FOUR_C_NAMESPACE_OPEN
 
 
 
-void Inpar::PoroScaTra::set_valid_parameters(std::map<std::string, Core::IO::InputSpec>& list)
+void PoroElastScaTra::set_valid_parameters(std::map<std::string, Core::IO::InputSpec>& list)
 {
   using namespace Core::IO::InputSpecBuilders;
