+++ src/poroelast_scatra/4C_poroelast_scatra_monolithic.hpp
@@ -13,9 +13,9 @@
  *----------------------------------------------------------------------*/
 #include "4C_config.hpp"
 
-#include "4C_inpar_poroscatra.hpp"
 #include "4C_linalg_mapextractor.hpp"
 #include "4C_poroelast_scatra_base.hpp"
+#include "4C_poroelast_scatra_input.hpp"
 #include "4C_utils_parameter_list.fwd.hpp"
 
 #include <Teuchos_Time.hpp>
